// Harken-v2-release2025.08/Harken-v2-release2025.08/packages/backend/src/utils/enums/AppraisalsEnum.ts
export enum ImagesEnum {
	HORIZONTAL = 'Horizontal',
	VERTICAL = 'Vertical',
}

export enum ImagesTitleEnum {
	SUB_PROPERTY = 'Sub. Property',
	SUB_PROPERTY_2 = 'Sub. Property 2',
	SUB_PROPERTY_3 = 'Sub. Property 3',
	SUB_PROPERTY_4 = 'Sub. Property 4',
	SUB_PROPERTY_5 = 'Sub. Property 5',
	SUB_PROPERTY_6 = 'Sub. Property 6',
	SUB_PROPERTY_7 = 'Sub. Property 7',
	SUB_PROPERTY_8 = 'Sub. Property 8',
	ADDITIONAL_IMAGE = 'Additional image',
	PHOTO_PAGE = 'Photo page',
}

export enum ImagesPageEnum {
	COVER_PHOTO = '1 - Cover Photo',
	TABLE_OF_CONTENTS = '2 - Table of Contents',
	EXECUTIVE_SUMMARY = '3 - Executive Summary',
	PROPERTY_SUMMARY = '4 - Property Summary',
	EXTRA_IMAGE = 'Extra image uploaded by user',
	COVER = 'cover',
	PHOTO = 'photo',
}